Add read only editor and memoize backlink rendering
